Messages can be seen on user created pages
A visitor can go to a url and view a page if they type in a password.
The implementation feels a bit hacky. Trying to avoid devise to keep
things light on dependencies (silly since I started this project with a
bunch of generic dependencies from `thoughtbot\suspenders` but sometimes
life is ironic like that.) Things do not have to make sense.

Permission class feels a bit silly since it just wraps the
`authenticate` method given by rails `secure_password` module.
Wanted to stay CRUD but look what happens.

Visiting a page twice is not allowed: the page is marked as seen. It
seems a little strange to do this all in the permissions controller.
Maybe only one controller necessary. Maybe devise would have been
better, but I still think it's heavy handed.