Allow Preview of markdown message while editing
The markdown javascript libraries do not parse with the same rules as
the redcarpet markdown gem. This leads to discrepancies between the
preview and the hidden page created.

The solution is to have an api for the message preview that is displayed
to the user while they edit the message. This could put heavy load on
the server (I paid for the cheapest option) so the JavaScript
`setTimeout` function is used. Users that stop typing for a second will
see the preview.
http://stackoverflow.com/questions/1909441/jquery-keyup-delay

One other feature spec modified so as not to depend on test order.
Databasecleaner gem could have been used instead, but this seems easier.
Instead of relying on `Page.count` to see if the page was created we
just check the `page.message`