Reset url to root when secret page is viewed
Adding this Javascript feels a bit hacky. Perhaps a better approach
would be to redirect rather than render the secrets page from the
permissions page. Then the url would display the secret page's url
(although this could be done with JS as well).

I avoided this because I did not want to create a session (or have the
user send back their password in the params) maybe I should force https
first anyway.