Use giphy api so users can search GIFs
https://github.com/Giphy/GiphyAPI

Shamelessly copied implementation at:
 http://codepen.io/blaketarter/pen/wBgWbV then simplified and modified.
This relies on a public key that is rate limited. If I want my own key I
would need to remove most of this JS and put the logic on the rails side
to avoid exposing the public key to the client. That would have to
double response time (client > my server > giphy > my server > client
versus the current (client > giphy > client)