Update to rails 5 and update production gems
* locking rails version at 5.0.0.1 (latest stable)
* Test/dev groups were updated in a previous commit:
* 8fbe04571716d6555ab1ab1f6d5d0dea823cbbb0

Fix 4 deprecation in config files:

* `config.serve_static_files` is deprecated
* Set cache control headers. Deprecation warning is not accurate since
this needs to be a hash not a string (maybe a dynamically created
deprecation warning, oops!)
http://blog.bigbinary.com/2015/10/31/rails-5-allows-setting-custom-http-headers-for-assets.html
* ActiveRecord::Base.raise_in_transactional_callbacks= is deprecated, has
no effect and will be removed without replacement.
* `redirect_to :back` is deprecated since HTTP_REFERER may not be
specified. Here we do not need to set a default since we know which page
the user came from in this case.