Improve and test ImageAdder JS object
Background on this object:
* This object is meant to handle the logic for when a user wants to use
a gif in their message. When they click the use button this object is
listening, and will save the url internally (imageMap array) and
generate a key for that Array that will appear in the message. The key
is a placeholder since with phones it is a bit annoying to see a long
url and try to copy and paste that.

What changed?
* This class had a few funky quirks including an unnecessary initialize
function, and a sad attempt at dependency injection.
* The initialize function bound an event handler and set the imageMap
(stateful component of this class) to an empty array. This does not need
to be reset as of yet. These two things can be handled up top in the
constructor class.
* A JS spec is added. To test this class interacting with the DOM we
need to build some elements, and stub the counter class currentImg
method (pulls the url for the GIF the user is looking at when they click
the 'use' button).

Quirk:
* The imageAdder in the test must be defined on window after the DOM is
built. Otherwise there is an imageAdder defined on window that did not
have event listeners bound at the right time.