Display optional fields when errors are present
The problem:
* Optional fields are hidden by default to make creating a note simpler.
However, if the user clicks more options, and then modifies these fields
but provides invalid input, it can be hard to see the errors after form
submission.
* Previously one would have to click the "more options" button again to
see errors for a duration outside the time limit, or an already taken
url_key.

The change:
* The hidden element that wraps these fields now has a dynamically
generated css class: "hidden or ""
* This ensures the initial state of these fields is visible if there are
errors on them. If there are no errors, the default is to have a
"hidden" class.

Implementation:
* This mix of presentation logic and error state seemed best outside of
the ActiveRecord model. A presenter is introduced and tested here.
Hidden fields are denoted in a class constant defined in the
PagePresenter.

A funny note on Big O:
* Doing an `any?` inside an `include?` may look suboptimal for
performance (n^2 for worst case) compared to set operators like `&`
http://ruby-doc.org/stdlib-2.3.3/libdoc/set/rdoc/Set.html#method-i-26
But when each array has under 5 max elements the overhead for the set
operators is way too large.