Refactor Preview and timer into separate objects
* The live preview uses a timer that generates a preview after every
keystroke followed by a one second delay.

* This change separates the delay logic and the preview generating logic
into separate objects. This allows for easier testing (a simple spec is
added for the preview functionality). The separate logic and testing may
be helpful for making the preview only appear when a button is pressed.
This saves many AJAX requests, reduces elements on the page that would
confuse a user, and matches the way Github allows markdown to be
previewed. This would requires some redesigning of this website.

Details:
* Some functions are broken down into smaller helper functions: e.g.
finalMessage. This preference for function over variable assignment can
help keep functions responsible for discreet bits of logic.
* A support/cleanup file is added to clear the html after each test.
This would cause duplicated page elements. I prefer to create the
elements used by the test at the top of each test file since this shows
clearly what needs to be on the page.

Later:
May follow up with stubbing some AJAX requests in the preview spec:
* https://github.com/jasmine/jasmine-ajax
* https://github.com/joneath/jasmine-ajax-mock