Refactor Gif scrolling logic into objects
* Implement a functions library for the functions that are somewhat
independent from this new class. These are already tested, but now the
file is renamed.

The new Giphy object communicates with the counter.
* The counter keeps track of current state (GIF img showing). This class
is purposefully decoupled from the DOM. It does not listen for clicks or
change the DOM. It can be told to increment with a direction/amount and
it will return it's state (which GIF image to show) through it's public
api currentImg.
* Giphy asks for the current img and tells the counter class to change
it's state. To test this interaction I thought it would be fun to mock
the counter class without using a mock library like Sinon.
* Other strange things done here include appending html in the test
rather than using a template.

Notes:
* The Giphy class has some distasteful code. It was basically copied and
pasted from a demo on how to use Giphy. It was never satisfactorily
rewritten. This will be tackled later. For now, the bigger OO
restructuring is the focus.