Allow Javascript tests to run in the browser
* Teaspoon tests can run in the browser or from the `rake teaspoon`
command line task.
* However, the after hook in `support/cleanup` was clearing the body
element on the page. This meant elements jasmine put in the body were
being wiped clean and causing errors in between tests.
* To solve this a new element is used that teaspoon would not know about
or use. That element can be safely cleared with `$el.html("")` after
each run. The element does need to be removed in a before hook.