Users can copy a link to clipboard with one click
* Using the `clipboard-rails` gem that wraps
https://github.com/zenorocha/clipboard.js it is easy to copy text to the
clipboard with most modern browsers
* The success page is restyled to incorporate this button. It looks a
lot better now.

* Adding a route helper so that `page_path` or `page_url` can be used in
the view file instead of concatenating strings. Some tests could also be
amended later because of this. I defined the `to_param` method on the
presenter rather than the page model itself. This keeps
view/presentation logic separate from persistence (rails uses to_param
to build the path to a resource. If this is not defined, id is the
default).

* The margin around buttons, textarea, inputs should be the same.
* Add margin on top of the body class for better look.