Make redirecting after several seconds optional
* By default pages would always redirect after 1-10 seconds.
Unfortunately, this limits the use of this website. Sending information
that the receiver wishes to copy and paste required quick copy/paste
skills.

* A redirect column is added to pages which determines whether the page
will include the Javascript that causes a redirect.

*** Challenges ***
* The stateful toggling functionality was already solved. When the
duration has errors, one would like that to not be hidden so the user
can see the errors. This leverages the presenter that was introduced to
toggle the optional fields.
* The preventDefault was being used to prevent the form submitting when
the "More Options" button was pressed. This button did not need to be
inside the form. It was removed, and then the previous button was
floated left to allow the two to appear next to each other.