Allow pages to be viewed multiple times (#94)
***If one wishes to post a vanishing to note to a group***
* This allows a page to be viewed multiple times. The pages expiration
controls when it is not visible. So a page set to expire in five minutes
could be viewed by multiple people (e.g. in a group text thread or slack
channel) before vanishing.

***Follow up***
* This could be a default option
* The timing options may need to change (30 seconds could be an option
even though that is pushing it with a free Heroku tier)
* Perhaps more instructions could be added here too.
* write a cron task to destroy the expired pages.

***Details***
* Instead of calling page.destroy! from the controller there is a new
`page#view!` method that decides whether to destroy the page. If the
boolean field `multiview` was set to `true` when the page was created
then the page will not be destroyed.
* The `unexpired` scope now prevents these pages from being destroyed.
* The cron tasks that used to destroy pages do not run on Heroku (this
was originally on digital ocean). This means the pages will sit around
for now.